results/rebuttal/run-20231111-034650_LLMatic-zero_shot_donnemartin-system-design-oop-bookclub_0_20-runs_log/bookclub/17/LLMatic/notification.py
class Notification:
	def __init__(self, database):
		self.database = database
		self.notifications = database.notifications

	def create_notification(self, id, data):
		self.database.insert(self.notifications, id, data)

	def get_notification(self, id):
		return self.database.get(self.notifications, id)

	def update_notification(self, id, data):
		if self.get_notification(id):
			self.database.insert(self.notifications, id, data)

	def delete_notification(self, id):
		self.database.delete(self.notifications, id)
